requirements.txt
# RLM-Trans 의존성

# Core
python-dotenv>=1.0.0
requests>=2.31.0

# LLM Providers
openai>=1.12.0
google-generativeai>=0.4.0

# Text Processing
langdetect>=1.0.9

# GUI
PyQt6>=6.6.0

# Optional: SRT parsing
pysrt>=1.1.2
